Consolidate delete budget endpoint into main controller
Merges the standalone delete budget controller into the main budget controller to reduce code duplication and improve maintainability. This consolidation keeps all budget-related endpoints in a single, cohesive controller while maintaining the same functionality and request/response handling.
