+++ src/main/java/com/code_factory/backend/budgeting/infrastructure/adapter/in/web/BudgetController.java
@@ -2,6 +2,7 @@ package com.code_factory.backend.budgeting.infrastructure.adapter.in.web;
 
 import com.code_factory.backend.budgeting.application.port.in.CreateBudgetCommand;
 import com.code_factory.backend.budgeting.application.port.in.CreateBudgetUseCase;
+import com.code_factory.backend.budgeting.application.port.in.DeleteBudgetUseCase;
 import com.code_factory.backend.budgeting.application.port.in.GetBudgetProgressUseCase;
 import com.code_factory.backend.budgeting.application.port.in.GetBudgetSummaryUseCase;
 import com.code_factory.backend.budgeting.application.port.in.ListBudgetsUseCase;
@@ -35,6 +36,7 @@ public class BudgetController {
     private final UpdateBudgetUseCase updateBudgetUseCase; 
     private final GetBudgetSummaryUseCase getBudgetSummaryUseCase;
     private final GetBudgetProgressUseCase getBudgetProgressUseCase;
+    private final DeleteBudgetUseCase deleteBudgetUseCase;
 
     @PostMapping
     public ResponseEntity<BudgetResponse> createBudget(@Valid @RequestBody CreateBudgetRequest request) {
@@ -156,4 +158,22 @@ public class BudgetController {
 
         return ResponseEntity.ok(response);
     }
+
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> deleteBudget(@PathVariable UUID id) {
+
+        boolean deleted = deleteBudgetUseCase.deleteById(id);
+
+        if (!deleted) {
+            return ResponseEntity
+                    .badRequest()
+                    .body("No existe un presupuesto para eliminar");
+        }
+
+        return ResponseEntity
+                .ok("Presupuesto eliminado correctamente");
+    }
+
+
 }
